requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas
scikit-learn

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow
s3fs

# utilities
six>=1.14
joblib
memoized-property
termcolor

# Trick to install the version of Tensorflow depending on your processor: darwin == Mac, ARM == M1
# tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
# tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
# tensorflow==2.10.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips

#local tensorflow M1 setup - NOT FOR PRODUCTION!
tensorflow-macos==2.8.0
tensorflow-metal==0.4.0

# Web App
# Streamlit etc
